Versie 3 divide method

+++ src/main/java/Calculator.java
@@ -12,7 +12,7 @@ public class Calculator {
         return a * b; // nu goed genoeg?
     }
 
-    public int divide (int a, int b) {
+    public double divide (double  a, double  b) {
         return a / b;
 
     }
